milestone_4/requirements.txt
streamlit
pandas
numpy
requests
python-dotenv
nltk
newsdataapi
google-generativeai
prophet
plotly
wordcloud
pillow
